+++ src/utils/autoprefix.js
@@ -2,10 +2,9 @@
 import camelizeStyleName from 'fbjs/lib/camelizeStyleName'
 import hyphenateStyleName from 'fbjs/lib/hyphenateStyleName'
 
-// eslint-disable-next-line import/no-duplicates
-import { autoprefix } from 'glamor/lib/autoprefix'
-// eslint-disable-next-line import/no-duplicates
-import typeof { Container } from 'glamor/lib/autoprefix'
+// eslint-disable-next-line
+import prefixAll from 'inline-style-prefixer/static'
+import type Container from '../vendor/postcss/container'
 
 export default (root: Container) => {
   root.walkDecls(decl => {
@@ -13,11 +12,15 @@ export default (root: Container) => {
     if (decl.prop.startsWith('--')) return
 
     const objStyle = { [camelizeStyleName(decl.prop)]: decl.value }
-    const prefixed = autoprefix(objStyle)
+    const prefixed = prefixAll(objStyle)
     Object.keys(prefixed).reverse().forEach(newProp => {
-      decl.cloneBefore({
-        prop: hyphenateStyleName(newProp),
-        value: prefixed[newProp],
+      const newVals = prefixed[newProp]
+      const newValArray = Array.isArray(newVals) ? newVals : [newVals]
+      newValArray.forEach(newVal => {
+        decl.cloneBefore({
+          prop: hyphenateStyleName(newProp),
+          value: newVal,
+        })
       })
     })
     decl.remove()