+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. If a contri
 ### Changed
 
 - Converted Object.assign to spread operator, thanks to [@thisguychris][https://github.com/hisguychris]. (see [#196](https://github.com/styled-components/styled-components/pull/201))
+- Switched to using [inline-style-prefixer](https://github.com/rofrischmann/inline-style-prefixer) for our autoprefixing needs.
 
 ## [v1.0.11] - 2016-11-14